backend/requirements.txt
# Django
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.4

# Rate limiting
django-ratelimit==4.1.0

# CORS
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# URL shortening
python-slugify==8.0.4

# QR Code generation
segno==1.6.1

# Production server
gunicorn==22.0.0
whitenoise==6.6.0

# Analytics
django-filter==24.2

# Geolocation
geoip2==4.7.0
requests==2.31.0

# Documentation
drf-spectacular==0.27.2

# Monitoring (optional)
django-prometheus==2.3.1

# Testing
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
